TASK: extract Sagas into an external package
