+++ packages/neos-ui-extensibility/src/shims/neosProjectPackages/neos-ui-sagas/index.js
@@ -1,0 +1,3 @@
+import readFromConsumerApi from '../../../readFromConsumerApi';
+
+module.exports = readFromConsumerApi('NeosProjectPackages')().NeosUiSagas;
