+++ packages/neos-ui-sagas/.babelrc
@@ -1,0 +1,3 @@
+{
+    "presets": ["@neos-project/babel-preset-neos-ui"]
+}
