+++ packages/neos-ui-sagas/package.json
@@ -1,0 +1,31 @@
+{
+  "name": "@neos-project/neos-ui-sagas",
+  "version": "1.3.4",
+  "description": "Redux Sagas implementation for the Neos CMS UI",
+  "private": true,
+  "main": "src/index.js",
+  "scripts": {
+    "prebuild": "check-dependencies && yarn clean",
+    "test": "yarn jest -- -w 2 --coverage",
+    "test:watch": "yarn jest -- --watch",
+    "build": "exit 0",
+    "build:watch": "exit 0",
+    "clean": "rimraf ./lib ./dist",
+    "lint": "eslint src",
+    "jest": "NODE_ENV=test jest"
+  },
+  "devDependencies": {
+    "@neos-project/babel-preset-neos-ui": "1.3.4",
+    "@neos-project/build-essentials": "1.3.4",
+    "@neos-project/jest-preset-neos-ui": "1.3.4"
+  },
+  "dependencies": {
+    "@neos-project/neos-ui-extensibility": "1.3.4",
+    "@neos-project/neos-ui-validators": "1.3.4",
+    "@neos-project/utils-redux": "1.3.4"
+  },
+  "license": "GNU GPLv3",
+  "jest": {
+    "preset": "@neos-project/jest-preset-neos-ui"
+  }
+}
