+++ packages/neos-ui-sagas/src/CR/NodeOperations/copyAndPasteNode.js
