+++ packages/neos-ui-sagas/src/CR/NodeOperations/cutAndPasteNode.js
