+++ packages/neos-ui-sagas/src/CR/NodeOperations/determineInsertMode.js
