+++ packages/neos-ui-sagas/src/CR/NodeOperations/helpers.js
