+++ packages/neos-ui-sagas/src/CR/NodeOperations/hideNode.js
