+++ packages/neos-ui-sagas/src/CR/NodeOperations/removeNodeIfConfirmed.js
