+++ packages/neos-ui/package.json
@@ -36,6 +36,7 @@
     "@neos-project/neos-ui-i18n": "1.3.4",
     "@neos-project/neos-ui-inspector": "1.3.4",
     "@neos-project/neos-ui-redux-store": "1.3.4",
+    "@neos-project/neos-ui-sagas": "1.3.4",
     "@neos-project/neos-ui-validators": "1.3.4",
     "@neos-project/neos-ui-views": "1.3.4",
     "@neos-project/positional-array-sorter": "1.3.4",
