+++ packages/neos-ui/src/manifest.js
@@ -458,6 +458,5 @@ manifest('main', {}, globalRegistry => {
 
 require('./manifest.containers');
 require('./manifest.dataloaders');
-require('./manifest.sagas');
 require('./manifest.reducer');
 require('./manifest.hotkeys');
